fixing the create product form and seprating the concerns into new hooks to fetch the product and admin id
